refactor(domain): package로 묶음
