+++ src/main/java/blackjack/domain/card/Shape.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public enum Shape {
     SPACE("스페이스"),
