+++ src/main/java/blackjack/domain/card/Value.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public enum Value {
     ACE(11),
