+++ src/main/java/blackjack/domain/result/Answer.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
 import java.util.Arrays;
 
@@ -16,7 +16,7 @@ public enum Answer {
         return Arrays.stream(values())
                 .filter(value -> value.answer.equals(answer))
                 .findFirst()
-                .orElseThrow(()-> new IllegalArgumentException("y, n로만 대답할 수 있습니다."));
+                .orElseThrow(() -> new IllegalArgumentException("y, n로만 대답할 수 있습니다."));
     }
 
     public static boolean isYes(Answer answer) {
