+++ src/main/java/blackjack/domain/user/Name.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.user;
 
 import java.util.regex.Pattern;
