+++ src/main/java/blackjack/domain/user/User.java
@@ -1,4 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.user;
+
+import blackjack.domain.card.Deck;
 
 import java.util.ArrayList;
