+++ src/test/java/blackjack/domain/result/ResultBoardTest.java
@@ -1,5 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.result;
 
+import blackjack.domain.user.Dealer;
+import blackjack.domain.user.Players;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
