+++ src/test/java/blackjack/domain/user/DealerTest.java
@@ -1,5 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.user;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Shape;
+import blackjack.domain.card.Value;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
